+++ JsonSubTypes.Tests/Properties/AssemblyInfo.cs
@@ -1,19 +1,6 @@
-using System.Reflection;
+﻿using System.Reflection;
 using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("JsonSubTypes.Tests")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("JsonSubTypes.Tests")]
-[assembly: AssemblyCopyright("Copyright ©  2017")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
 
 [assembly: ComVisible(false)]
 
 [assembly: Guid("dfe69232-1753-4fad-acf3-e8a9bbf434f1")]
-
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]