+++ JsonSubTypes/JsonSubtypes.cs
@@ -221,7 +221,7 @@ namespace JsonSubTypes
             return jObjectReader;
         }
 
-        private Type GetType(JObject jObject, Type parentType)
+        private Type ResolveType(JObject jObject, Type parentType, JsonSerializer serializer)
         {
             Type resolvedType;
             if (JsonDiscriminatorPropertyName == null)
@@ -230,7 +230,7 @@ namespace JsonSubTypes
             }
             else
             {
-                resolvedType = GetTypeFromDiscriminatorValue(jObject, parentType);
+                resolvedType = GetTypeFromDiscriminatorValue(jObject, parentType, serializer);
             }
 
             return resolvedType ?? GetFallbackSubType(parentType);
@@ -241,11 +241,17 @@ namespace JsonSubTypes
             Type targetType = parentType;
             JsonSubtypes lastTypeResolver = null;
             JsonSubtypes currentTypeResolver = this;
+            var visitedTypes = new HashSet<Type> { targetType };
 
             var jsonConverterCollection = serializer.Converters.OfType<JsonSubtypesByDiscriminatorValueConverter>().ToList();
             while (currentTypeResolver != null && currentTypeResolver != lastTypeResolver)
             {
-                targetType = currentTypeResolver.GetType(jObject, targetType);
+                targetType = currentTypeResolver.ResolveType(jObject, targetType, serializer);
+                if (!visitedTypes.Add(targetType))
+                {
+                    break;
+                }
+
                 lastTypeResolver = currentTypeResolver;
                 jsonConverterCollection = jsonConverterCollection.Where(c => c != currentTypeResolver).ToList();
                 currentTypeResolver = GetTypeResolver(ToTypeInfo(targetType), jsonConverterCollection);
@@ -271,7 +277,7 @@ namespace JsonSubTypes
                 .FirstOrDefault(c => c.CanConvert(ToType(targetType)));
         }
 
-        private Type GetTypeByPropertyPresence(IDictionary<string, JToken> jObject, Type parentType)
+        private Type GetTypeByPropertyPresence(JObject jObject, Type parentType)
         {
             var knownSubTypeAttributes = GetTypesByPropertyPresence(parentType);
 
@@ -280,6 +286,12 @@ namespace JsonSubTypes
                 {
                     if (TryGetValueInJson(jObject, knownType.Key, out JToken _))
                         return knownType.Value;
+
+                    var token = jObject.SelectToken(knownType.Key);
+                    if (token != null)
+                    {
+                        return knownType.Value;
+                    }
 
                     return null;
                 })
@@ -292,15 +304,22 @@ namespace JsonSubTypes
                 .ToDictionary(a => a.PropertyName, a => a.SubType);
         }
 
-        private Type GetTypeFromDiscriminatorValue(IDictionary<string, JToken> jObject, Type parentType)
+        private Type GetTypeFromDiscriminatorValue(JObject jObject, Type parentType, JsonSerializer serializer)
         {
             if (!TryGetValueInJson(jObject, JsonDiscriminatorPropertyName, out var discriminatorValue))
+            {
+                discriminatorValue = jObject.SelectToken(JsonDiscriminatorPropertyName);
+            }
+
+            if (discriminatorValue == null)
+            {
                 return null;
+            }
 
             var typeMapping = GetSubTypeMapping(parentType);
             if (typeMapping.Entries().Any())
             {
-                return GetTypeFromMapping(typeMapping, discriminatorValue);
+                return GetTypeFromMapping(typeMapping, discriminatorValue, serializer);
             }
 
             return GetTypeByName(discriminatorValue.Value<string>(), ToTypeInfo(parentType));
@@ -353,7 +372,7 @@ namespace JsonSubTypes
             return null;
         }
 
-        private static Type GetTypeFromMapping(NullableDictionary<object, Type> typeMapping, JToken discriminatorToken)
+        private static Type GetTypeFromMapping(NullableDictionary<object, Type> typeMapping, JToken discriminatorToken, JsonSerializer serializer)
         {
             if (discriminatorToken.Type == JTokenType.Null)
             {
@@ -366,7 +385,7 @@ namespace JsonSubTypes
             if (key != null)
             {
                 var targetLookupValueType = key.GetType();
-                var lookupValue = discriminatorToken.ToObject(targetLookupValueType);
+                var lookupValue = discriminatorToken.ToObject(targetLookupValueType, serializer);
 
                 if (typeMapping.TryGetValue(lookupValue, out Type targetType))
                 {
@@ -427,7 +446,7 @@ namespace JsonSubTypes
 #endif
         }
 
-        private static TypeInfo ToTypeInfo(Type type)
+        internal static TypeInfo ToTypeInfo(Type type)
         {
 #if (NET35 || NET40)
             return type;
@@ -436,7 +455,7 @@ namespace JsonSubTypes
 #endif
         }
 
-        private static Type ToType(TypeInfo typeInfo)
+        internal static Type ToType(TypeInfo typeInfo)
         {
 #if (NET35 || NET40)
             return typeInfo;